+++ tests/testthat/test-npn-phenophases.R
@@ -86,7 +86,7 @@ test_that("npn_get_phenophases_for_taxon works", {
   skip_if_not(check_service(), "Service is down")
 
   vcr::use_cassette("npn_get_phenophases_for_taxon_1", {
-    pp <- npn_get_phenophases_for_taxon(class_ids=5, date="2018-05-05")
+    pp <- npn_get_phenophases_for_taxon(class_ids = 5, date = "2018-05-05")
   })
 
   expect_s3_class(pp, "data.frame")
@@ -94,7 +94,7 @@ test_that("npn_get_phenophases_for_taxon works", {
   expect_equal(nrow(pp), 21)
 
   vcr::use_cassette("npn_get_phenophases_for_taxon_2", {
-    pp <- npn_get_phenophases_for_taxon(class_ids=c(5,6), date="2018-05-05")
+    pp <- npn_get_phenophases_for_taxon(class_ids = c(5, 6), date = "2018-05-05")
   })
 
   expect_s3_class(pp, "data.frame")
@@ -102,7 +102,7 @@ test_that("npn_get_phenophases_for_taxon works", {
   expect_equal(nrow(pp), 28)
 
   vcr::use_cassette("npn_get_phenophases_for_taxon_3", {
-    pp <- npn_get_phenophases_for_taxon(family_ids=c(267,268),date="2018-05-05")
+    pp <- npn_get_phenophases_for_taxon(family_ids = c(267, 268), date = "2018-05-05")
   })
 
   expect_s3_class(pp, "data.frame")
@@ -110,28 +110,10 @@ test_that("npn_get_phenophases_for_taxon works", {
   expect_equal(nrow(pp), 24)
 
   vcr::use_cassette("npn_get_phenophases_for_taxon_4", {
-    pp <- npn_get_phenophases_for_taxon(order_ids=c(74,75), date="2018-05-05", return_all = 0)
-  })
-
-  expect_s3_class(pp, "data.frame")
-  expect_type(pp$order_name, "character")
-  expect_equal(nrow(pp), 21)
-
-
-  vcr::use_cassette("npn_get_phenophases_for_taxon_5", {
-    pp <- npn_get_phenophases_for_taxon(order_ids=c(74,75), return_all = 1)
+    pp <- npn_get_phenophases_for_taxon(order_ids = c(74, 75), date = "all")
   })
 
   expect_s3_class(pp, "data.frame")
   expect_type(pp$order_name, "character")
   expect_equal(nrow(pp), 155)
-
-  skip("unclear if this last one is supposed to work or error")
-  vcr::use_cassette("npn_get_phenophases_for_taxon_6", {
-    pp <- npn_get_phenophases_for_taxon(order_ids=c(74,75), return_all = TRUE)
-  })
-
-  expect_s3_class(pp, "data.frame")
-  expect_type(pp$order_name, "character")
-  expect_equal(nrow(pp), 2)
 })